// frontend/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        RoomSnapshotResponse: {
            /** Game Id */
            game_id?: string | null;
            /** Last Game Id */
            last_game_id?: string | null;
            /** Max Participants */
            max_participants: number;
            /** Minimum Ready */
            minimum_ready: number;
            /** Name */
            name: string;
            /** Owner Id */
            owner_id: string | null;
            /** Participants */
            participants: components["schemas"]["RoomParticipantResponse"][];
            /** Password Required */
            password_required: boolean;
            /**
             * Replayed
             * @default false
             */
            replayed: boolean;
            /** Room Id */
            room_id: string;
            /** State Version */
            state_version: number;
            status: components["schemas"]["RoomStatus"];
            visibility: components["schemas"]["RoomVisibility"];
            /** Vote Seconds */
            vote_seconds: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}